optional name for timestamp in prediction_times_df
Name for timestamp column was optional for prediction_times_df, but the function didn't run if the name wasn't '"timestamp". Tried to alter to manual "timestamp" names to self.timestamp_col_name, but then the code following the merge with quarantine_df broke, since the suffixes weren't added. So changed the name of the timestamp in quarantine_df in the start of the script.