dev: add quick PR using testmon (#389)
This adds an entrypoint for running tests using https://testmon.org/

> pytest-testmon is a pytest plugin which selects and executes only
tests you need to run. It does this by collecting dependencies between
tests and all executed code (internally using Coverage.py) and comparing
the dependencies against changes. testmon updates its database on each
test execution, so it works independently of version control.

This dramatically decreases our test run times, especially for a
monorepo.

Potential downsides are false negatives. To avoid this, we still run the
full test suite in CI.

I've added it as a separate entrypoint so we can test it, before
suggesting it as an integration in everyone's workflow. This is tracked
in #390, so we don't forget.

Since it uses code coverage, it won't catch if you have changed an
external file. E.g. if a .cfg file changes, any tests using the .cfg
file won't run. I think that is fine as long as we run the full test
suite remotely. Alternatively, we could run both locally. First
testmon-tests (to get fast feedback), then all tests (to get
comprehensive feedback).

@KennethEnevoldsen, @HLasse, what do you guys think? :-)